// src/app/product/models/product-item.model.ts
export interface ProductItemModel {
  Id?: any;
  Name?: any;
  ProductName?: any;
  ShortDescription?: any;
  FullDescription?: any;
  MetaDescription?: any;
  FarahooshComment?: any;
  MetaKeywords?: any;
  MetaTitle?: any;
  BrandId?: any;

  BrandName?: any;
  Brand?: any;
  CategoryName?: any;
  Sku?: any;
  ManufacturerPartNumber?: any;
  Gtin?: any;
  HSCodeId?: any;
  OtherDetails?: any;
  RequiredProductIds?: any;
  VisibleIndividually?: boolean;
  AllowCustomerReviews?: boolean;
  ApprovedRatingSum?: number;
  ApprovedTotalReviews?: number;
  NotApprovedRatingSum?: number;
  NotApprovedTotalReviews?: number;
  ShowOnHomepage?: boolean;
  IsGiftCard?: boolean;
  GiftCardTypeId?: number;
  OverriddenGiftCardAmount?: any;
  RequireOtherProducts?: boolean;
  AutomaticallyAddRequiredProducts?: boolean;
  HasSampleDownload?: boolean;
  HTSNumber?: string;
  DownloadExpirationDays?: number;
  UserAgreementText?: any;
  IsRecurring?: boolean;
  IsRental?: boolean;
  ShowNotifyAfterExpireRentalDate?: boolean;
  HasDownloadFile?: boolean;
  IsShipEnabled?: boolean;
  DeliveryDateId?: number;
  TaxCategory?: number;
  TaxCategoryId?: number;
  DisplayStockAvailability?: boolean;
  DisplayStockQuantity?: boolean;
  NotifyAdminForQuantityBelow?: boolean;
  LowStockActivityId?: number;
  NotReturnable?: boolean;
  DisableWishlistButton?: boolean;
  HasProductTag?: boolean;
  AvailableStartDateTimeUtc?: any;
  AvailableEndDateTimeUtc?: any;
  DisplayOrder?: number;
  AvailableForPreOrder?: boolean;
  PreOrderAvailabilityStartDateTimeUtc?: any;
  Published?: boolean;
  Deleted?: boolean;
  currentName?: string;
  currentShortDescription?: string;
  currentFullDescription?: string;
  currentBrand?: string;
  currentCategory?: string;
  CreateDateUTC?: string;
  Status?: any;
  StatusComment?: any;
  languages?: string[];
  ProductGuid?: string;
  ProductionCapacity?: any;
  VendorFAQ?: any;
  PackagingAndShipping?: any;
  FAQ?: any;
  ProductionLocationList?: any;
  SendLocationList?: any;
  TaxCategoryName?: string;
  Tag?: any;
  pkgHeight: any;
  pkgLength: any;
  pkgNature: any;
  pkgType: any;
  pkgWeigth: any;
  pkgWidth: any;
  pkhWeigthUnit: any;
  ReadyToShip?: boolean;
  IsCustomizable?: boolean;
}

export interface ProductDetail {
  Detail: ProductItemModel;
  Attribut: any[];
  Category: any[];
  Tag: any;
  ProductTransportation: any[];
}